simplex-outreach/requirements.txt
streamlit==1.31.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
dnspython==2.5.0
pandas==2.0.3
